+++ app/gui2/src/components/GraphEditor/NodeWidget.vue
@@ -1,6 +1,9 @@
 <script setup lang="ts">
-import type { PortId } from '@/providers/portInfo'
-import { injectWidgetRegistry, type WidgetInput } from '@/providers/widgetRegistry'
+import {
+  injectWidgetRegistry,
+  type UpdatePayload,
+  type WidgetInput,
+} from '@/providers/widgetRegistry'
 import { injectWidgetTree } from '@/providers/widgetTree'
 import {
   injectWidgetUsageInfo,
@@ -13,6 +16,7 @@ import { computed, proxyRefs } from 'vue'
 const props = defineProps<{
   input: WidgetInput
   nest?: boolean
+  allowEmpty?: boolean
   /**
    * A function that intercepts and handles a value update emitted by this widget. When it returns
    * `false`, the update continues to be propagated to the parent widget. When it returns `true`,
@@ -24,7 +28,7 @@ defineOptions({
   inheritAttrs: false,
 })
 
-type UpdateHandler = (value: unknown, origin: PortId) => boolean
+type UpdateHandler = (update: UpdatePayload) => boolean
 
 const registry = injectWidgetRegistry()
 const tree = injectWidgetTree()
@@ -52,9 +56,9 @@ const updateHandler = computed(() => {
     parentUsageInfo?.updateHandler ?? (() => console.log('Missing update handler'))
   if (props.onUpdate != null) {
     const localHandler = props.onUpdate
-    return (value: unknown, origin: PortId) => {
-      const handled = localHandler(value, origin)
-      if (!handled) nextHandler(value, origin)
+    return (payload: UpdatePayload) => {
+      const handled = localHandler(payload)
+      if (!handled) nextHandler(payload)
     }
   }
   return nextHandler
@@ -97,7 +101,7 @@ const spanStart = computed(() => {
     @update="updateHandler"
   />
   <span
-    v-else
+    v-else-if="!props.allowEmpty"
     :title="`No matching widget for input: ${
       Object.getPrototypeOf(props.input)?.constructor?.name ?? JSON.stringify(props.input)
     }`"