+++ app/gui2/src/components/GraphEditor/widgets/WidgetCheckbox.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import CheckboxWidget from '@/components/widgets/CheckboxWidget.vue'
-import { AnyWidget, Score, defineWidget, widgetProps } from '@/providers/widgetRegistry'
+import { Score, WidgetInput, defineWidget, widgetProps } from '@/providers/widgetRegistry'
 import { Ast } from '@/util/ast'
 import { computed } from 'vue'
 
@@ -8,13 +8,20 @@ const props = defineProps(widgetProps(widgetDefinition))
 
 const value = computed({
   get() {
-    return props.input.ast?.code().endsWith('True') ?? false
+    return WidgetInput.valueRepr(props.input)?.endsWith('True') ?? false
   },
   set(value) {
-    if (props.input.ast == null) return // TODO[ao] set value on placeholder here.
-    const node = getRawBoolNode(props.input.ast)
-    if (node != null) {
-      props.onUpdate(value ? 'True' : 'False', node.exprId)
+    if (props.input.value instanceof Ast.Ast) {
+      const node = getRawBoolNode(props.input.value)
+      if (node != null) {
+        props.onUpdate({ type: 'set', value: value ? 'True' : 'False', origin: node.exprId })
+      }
+    } else {
+      props.onUpdate({
+        type: 'set',
+        value: value ? 'Boolean.True' : 'Boolean.False',
+        origin: props.input.portId,
+      })
     }
   },
 })
@@ -30,13 +37,14 @@ function getRawBoolNode(ast: Ast.Ast) {
   return null
 }
 
-export const widgetDefinition = defineWidget(AnyWidget, {
-  priority: 10,
+export const widgetDefinition = defineWidget(WidgetInput.isAstOrPlaceholder, {
+  priority: 1001,
   score: (props) => {
-    if (props.input.ast == null)
-      return props.input.argInfo?.reprType === 'Standard.Base.Bool' ? Score.Good : Score.Mismatch
-    if (getRawBoolNode(props.input.ast) != null) return Score.Perfect
-    return Score.Mismatch
+    if (props.input.value instanceof Ast.Ast && getRawBoolNode(props.input.value) != null)
+      return Score.Perfect
+    return props.input.expectedType === 'Standard.Base.Data.Boolean.Boolean'
+      ? Score.Good
+      : Score.Mismatch
   },
 })
 </script>