+++ lib/scala/project-manager/src/test/scala/org/enso/projectmanager/protocol/ProjectOpenMissingComponentsSpec.scala
@@ -1,10 +1,12 @@
 package org.enso.projectmanager.protocol
 import nl.gn0s1s.bump.SemVer
+import org.enso.logger.ReportLogsOnFailure
 import org.enso.runtimeversionmanager.test.OverrideTestVersionSuite
 
 class ProjectOpenMissingComponentsSpec
     extends ProjectOpenSpecBase
-    with OverrideTestVersionSuite {
+    with OverrideTestVersionSuite
+    with ReportLogsOnFailure {
 
   override val testVersion: SemVer = defaultVersion