+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/Storage.java
@@ -415,7 +415,20 @@ public abstract class Storage<T> {
   public Storage<?> fillMissing(
       Value arg, StorageType commonType, ProblemAggregator problemAggregator) {
     Builder builder = Builder.getForType(commonType, size(), problemAggregator);
-    return fillMissingHelper(arg, builder);
+    Object convertedFallback = Polyglot_Utils.convertPolyglotValue(arg);
+    Context context = Context.getCurrent();
+    for (int i = 0; i < size(); i++) {
+      Object it = getItemBoxed(i);
+      if (it == null) {
+        builder.appendNoGrow(convertedFallback);
+      } else {
+        builder.appendNoGrow(it);
+      }
+
+      context.safepoint();
+    }
+
+    return builder.seal();
   }
 
   /**
@@ -434,23 +447,6 @@ public abstract class Storage<T> {
         builder.appendNoGrow(other.getItemBoxed(i));
       } else {
         builder.appendNoGrow(getItemBoxed(i));
-      }
-
-      context.safepoint();
-    }
-
-    return builder.seal();
-  }
-
-  protected final Storage<?> fillMissingHelper(Value arg, Builder builder) {
-    Object convertedFallback = Polyglot_Utils.convertPolyglotValue(arg);
-    Context context = Context.getCurrent();
-    for (int i = 0; i < size(); i++) {
-      Object it = getItemBoxed(i);
-      if (it == null) {
-        builder.appendNoGrow(convertedFallback);
-      } else {
-        builder.appendNoGrow(it);
       }
 
       context.safepoint();